AI_Employee_Vault/scripts/requirements.txt
# AI Employee Bronze Tier - Python Dependencies

# File system monitoring (for filesystem_watcher.py)
watchdog>=4.0.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: For Gmail watcher (Silver tier)
# google-api-python-client>=2.0.0
# google-auth-httplib2>=0.2.0
# google-auth-oauthlib>=1.0.0

# Optional: For WhatsApp watcher (Silver tier)
# playwright>=1.40.0
